# frontend/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Telegram user profiles",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/channels": {
      "get": {
        "summary": "Get Channels",
        "operationId": "get_channels",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelsResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Save Channels List",
        "operationId": "save_channels_list",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChannelsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/collect": {
      "post": {
        "summary": "Start Collect",
        "operationId": "start_collect",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStarted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/collect/{job_id}/status": {
      "get": {
        "summary": "Collect Status",
        "operationId": "collect_status",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/collect/{job_id}/cancel": {
      "post": {
        "summary": "Cancel Collect",
        "operationId": "cancel_collect",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobCancelled"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{tg_id}": {
      "get": {
        "summary": "Get User",
        "operationId": "get_user",
        "parameters": [
          {
            "name": "tg_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tg Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{tg_id}/comments.txt": {
      "get": {
        "summary": "Export User Comments",
        "operationId": "export_user_comments",
        "parameters": [
          {
            "name": "tg_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tg Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{tg_id}/political-coords": {
      "post": {
        "summary": "Analyze Political",
        "operationId": "analyze_political",
        "parameters": [
          {
            "name": "tg_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tg Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PoliticalCoords"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{tg_id}/position-comparisons": {
      "get": {
        "summary": "Get Position Comparisons",
        "operationId": "get_position_comparisons",
        "parameters": [
          {
            "name": "tg_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tg Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionResults"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnalysisProgress": {
        "properties": {
          "state": {
            "type": "string",
            "enum": [
              "idle",
              "running",
              "done",
              "error",
              "interrupted",
              "provider_changed"
            ],
            "title": "State",
            "default": "idle"
          },
          "phase": {
            "type": "string",
            "title": "Phase",
            "default": ""
          },
          "processed_comments": {
            "type": "integer",
            "title": "Processed Comments",
            "default": 0
          },
          "total_comments": {
            "type": "integer",
            "title": "Total Comments",
            "default": 0
          },
          "cached_comments": {
            "type": "integer",
            "title": "Cached Comments",
            "default": 0
          },
          "rejected_comments": {
            "type": "integer",
            "title": "Rejected Comments",
            "default": 0
          },
          "rejection_reasons": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Rejection Reasons"
          },
          "processed_embeddings": {
            "type": "integer",
            "title": "Processed Embeddings",
            "default": 0
          },
          "total_embeddings": {
            "type": "integer",
            "title": "Total Embeddings",
            "default": 0
          },
          "cached_embeddings": {
            "type": "integer",
            "title": "Cached Embeddings",
            "default": 0
          },
          "processed_questions": {
            "type": "integer",
            "title": "Processed Questions",
            "default": 0
          },
          "total_questions": {
            "type": "integer",
            "title": "Total Questions",
            "default": 0
          },
          "extraction_requests": {
            "type": "integer",
            "title": "Extraction Requests",
            "default": 0
          },
          "split_batches": {
            "type": "integer",
            "title": "Split Batches",
            "default": 0
          },
          "activity": {
            "type": "string",
            "title": "Activity",
            "default": ""
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "processed_pairs": {
            "type": "integer",
            "title": "Processed Pairs",
            "default": 0
          },
          "total_pairs": {
            "type": "integer",
            "title": "Total Pairs",
            "default": 0
          },
          "processed_relations": {
            "type": "integer",
            "title": "Processed Relations",
            "default": 0
          },
          "total_relations": {
            "type": "integer",
            "title": "Total Relations",
            "default": 0
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "state",
          "phase",
          "processed_comments",
          "total_comments",
          "cached_comments",
          "rejected_comments",
          "rejection_reasons",
          "processed_embeddings",
          "total_embeddings",
          "cached_embeddings",
          "processed_questions",
          "total_questions",
          "extraction_requests",
          "split_batches",
          "activity",
          "updated_at",
          "processed_pairs",
          "total_pairs",
          "processed_relations",
          "total_relations",
          "error"
        ],
        "title": "AnalysisProgress"
      },
      "AxisCounts": {
        "properties": {
          "left_count": {
            "type": "integer",
            "title": "Left Count"
          },
          "right_count": {
            "type": "integer",
            "title": "Right Count"
          }
        },
        "type": "object",
        "required": [
          "left_count",
          "right_count"
        ],
        "title": "AxisCounts"
      },
...
      "ChannelStatus": {
        "properties": {
          "channel": {
            "type": "string",
            "title": "Channel"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "saved": {
            "type": "integer",
            "title": "Saved"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "channel",
          "status",
          "saved",
          "error"
        ],
        "title": "ChannelStatus"
      },
      "ChannelsRequest": {
        "properties": {
          "channels_text": {
            "type": "string",
            "title": "Channels Text"
          }
        },
        "type": "object",
        "required": [
          "channels_text"
        ],
        "title": "ChannelsRequest"
      },
      "ChannelsResponse": {
        "properties": {
          "channels": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Channels"
          }
        },
        "type": "object",
        "required": [
          "channels"
        ],
        "title": "ChannelsResponse"
      },
      "CollectRequest": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          }
        },
        "type": "object",
        "required": [
          "username"
        ],
        "title": "CollectRequest"
      },
...
      "DayCount": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "date",
          "count"
        ],
        "title": "DayCount"
      },
      "EmbeddingDetails": {
        "properties": {
          "model": {
            "type": "string",
            "title": "Model"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "dimensions": {
            "type": "integer",
            "title": "Dimensions"
          },
          "storage": {
            "type": "string",
            "title": "Storage"
          },
          "saved_vectors": {
            "type": "integer",
            "title": "Saved Vectors"
          }
        },
        "type": "object",
        "required": [
          "model",
          "version",
          "dimensions",
          "storage",
          "saved_vectors"
        ],
        "title": "EmbeddingDetails"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HourCount": {
        "properties": {
          "hour": {
            "type": "integer",
            "title": "Hour"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "hour",
          "count"
        ],
        "title": "HourCount"
      },
      "JobCancelled": {
        "properties": {
          "cancelled": {
            "type": "boolean",
            "title": "Cancelled"
          }
        },
        "type": "object",
        "required": [
          "cancelled"
        ],
        "title": "JobCancelled"
      },
      "JobStarted": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          }
        },
        "type": "object",
        "required": [
          "job_id"
        ],
        "title": "JobStarted"
      },
      "JobStatus": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "user_ref": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "string"
              }
            ],
            "title": "User Ref"
          },
          "total_channels": {
            "type": "integer",
            "title": "Total Channels"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "resolved": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResolvedUser"
              },
              {
                "type": "null"
              }
            ]
          },
          "channels": {
            "items": {
              "$ref": "#/components/schemas/ChannelStatus"
            },
            "type": "array",
            "title": "Channels"
          },
          "saved_total": {
            "type": "integer",
            "title": "Saved Total"
          },
          "tg_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tg Id"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "user_ref",
          "total_channels",
          "state",
          "resolved",
          "channels",
          "saved_total",
          "tg_id",
          "error"
        ],
        "title": "JobStatus"
      },
      "PoliticalCoords": {
        "properties": {
          "total_messages": {
            "type": "integer",
            "title": "Total Messages"
          },
          "signal_count": {
            "type": "integer",
            "title": "Signal Count"
          },
          "bars": {
            "type": "string",
            "title": "Bars"
          },
          "axes": {
            "additionalProperties": {
              "$ref": "#/components/schemas/AxisCounts"
            },
            "type": "object",
            "title": "Axes"
          }
        },
        "type": "object",
        "required": [
          "total_messages",
          "signal_count",
          "bars",
          "axes"
        ],
        "title": "PoliticalCoords"
      },
      "PositionResults": {
        "properties": {
          "version": {
            "type": "string",
            "title": "Version"
          },
          "needs_update": {
            "type": "boolean",
            "title": "Needs Update",
            "default": false
          },
          "incomplete": {
            "type": "boolean",
            "title": "Incomplete",
            "default": false
          },
          "progress": {
            "$ref": "#/components/schemas/AnalysisProgress"
          },
          "embeddings": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EmbeddingDetails"
              },
              {
                "type": "null"
              }
            ]
          },
          "method": {
            "type": "string",
            "enum": [
              "position_agreement",
              "text_similarity"
            ],
            "title": "Method",
            "default": "position_agreement"
          },
          "similar_authors": {
            "items": {
              "$ref": "#/components/schemas/SimilarAuthor"
            },
            "type": "array",
            "title": "Similar Authors"
          },
          "ranking": {
            "items": {
              "$ref": "#/components/schemas/Comparison"
            },
            "type": "array",
            "title": "Ranking"
          },
          "insufficient": {
            "items": {
              "$ref": "#/components/schemas/Comparison"
            },
            "type": "array",
            "title": "Insufficient"
          },
          "partial_results": {
            "items": {
              "$ref": "#/components/schemas/Comparison"
            },
            "type": "array",
            "title": "Partial Results"
          }
        },
        "type": "object",
        "required": [
          "version",
          "needs_update",
          "incomplete",
          "progress",
          "embeddings",
          "method",
          "similar_authors",
          "ranking",
          "insufficient",
          "partial_results"
        ],
        "title": "PositionResults"
      },
      "Profile": {
        "properties": {
          "tg_id": {
            "type": "integer",
            "title": "Tg Id"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "display_username": {
            "type": "string",
            "title": "Display Username"
          },
          "total_messages": {
            "type": "integer",
            "title": "Total Messages"
          },
          "channel_count": {
            "type": "integer",
            "title": "Channel Count"
          },
          "channels": {
            "items": {
              "$ref": "#/components/schemas/ChannelShare"
            },
            "type": "array",
            "title": "Channels"
          }
        },
        "type": "object",
        "required": [
          "tg_id",
          "username",
          "display_name",
          "display_username",
          "total_messages",
          "channel_count",
          "channels"
        ],
        "title": "Profile"
      },
...
      "ResolvedUser": {
        "properties": {
          "tg_id": {
            "type": "integer",
            "title": "Tg Id"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          }
        },
        "type": "object",
        "required": [
          "tg_id",
          "username",
          "display_name"
        ],
        "title": "ResolvedUser"
      },
      "SimilarAuthor": {
        "properties": {
          "tg_id": {
            "type": "integer",
            "title": "Tg Id"
          },
          "display_username": {
            "type": "string",
            "title": "Display Username"
          },
          "similarity": {
            "type": "number",
            "title": "Similarity"
          },
          "left_comments": {
            "type": "integer",
            "title": "Left Comments"
          },
          "right_comments": {
            "type": "integer",
            "title": "Right Comments"
          },
          "examples": {
            "items": {
              "$ref": "#/components/schemas/TextMatch"
            },
            "type": "array",
            "title": "Examples"
          }
        },
        "type": "object",
        "required": [
          "tg_id",
          "display_username",
          "similarity",
          "left_comments",
          "right_comments",
          "examples"
        ],
        "title": "SimilarAuthor"
      },
      "TextMatch": {
        "properties": {
          "similarity": {
            "type": "number",
            "title": "Similarity"
          },
          "left": {
            "$ref": "#/components/schemas/Evidence"
          },
          "right": {
            "$ref": "#/components/schemas/Evidence"
          }
        },
        "type": "object",
        "required": [
          "similarity",
          "left",
          "right"
        ],
        "title": "TextMatch"
      },
      "UserDetail": {
        "properties": {
          "profile": {
            "$ref": "#/components/schemas/Profile"
          },
          "hourly_activity": {
            "items": {
              "$ref": "#/components/schemas/HourCount"
            },
            "type": "array",
            "title": "Hourly Activity"
          },
          "daily_activity": {
            "items": {
              "$ref": "#/components/schemas/DayCount"
            },
            "type": "array",
            "title": "Daily Activity"
          },
          "weekly_activity": {
            "items": {
              "$ref": "#/components/schemas/WeekHourCount"
            },
            "type": "array",
            "title": "Weekly Activity"
          }
        },
        "type": "object",
        "required": [
          "profile",
          "hourly_activity",
          "daily_activity",
          "weekly_activity"
        ],
        "title": "UserDetail"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WeekHourCount": {
        "properties": {
          "weekday": {
            "type": "integer",
            "title": "Weekday"
          },
          "hour": {
            "type": "integer",
            "title": "Hour"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "weekday",
          "hour",
          "count"
        ],
        "title": "WeekHourCount"
      }
    }
  }
}